feat(web): extract AI source detection into ai-sources-lib
Move isomorphic referrer and crawler matchers into a pure lib module with wrapper parity tests while keeping the public ai-sources surface.
